requirements.txt
Django==4.1.1
psycopg2==2.9.3
xhtml2pdf==0.2.8
Pillow==9.2.0
django-widget-tweaks==1.4.12
django-crum==0.7.9
